use float custom vertex attrs

+++ src/render/antic.vert
@@ -4,11 +4,11 @@ precision highp int;
 
 in vec3 Vertex_Position;
 in vec2 Vertex_Uv;
-in uvec4 Vertex_Custom;
+in vec4 Vertex_Custom;
 
 out vec3 v_Position;
 out vec2 v_Uv;
-flat out uvec4 v_Custom;
+flat out vec4 v_Custom;
 
 layout(std140) uniform Camera {
     mat4 ViewProj;
